Implement functionality where we prevent making polls to server when an error has occured
So initially it would happen that not only was the system making polls, we allowed the user to "try again" what this did -- was duplicate the announcements locally. When the user refreshes now, we cancel and re-initialzie the timer to effectively skip that particular poll.

Other than this, we now also distinguish between a primary error, which may occur when loading the feed itself, and a secondary error which is less intrusive and may occur when updating the feed